planner interface refactor
